TST Add new test file for testing depmtx
